pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["ingestor_reader"]

[tool.setuptools.package-data]
ingestor_reader = ["**/*"]

[project]
name = "ingestor-reader"
version = "0.1.0"
description = "Minimal config-first ETL service"
requires-python = ">=3.11"
dependencies = [
    "boto3>=1.28.0",
    "botocore>=1.31.0",
    "pyarrow>=14.0.0",
    "pandas>=2.0.0",
    "pydantic>=2.0.0",
    "typer>=0.9.0",
    "openpyxl>=3.1.0",
    "pyyaml>=6.0.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "moto>=5.0.0",
    "pytest-cov>=4.1.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"

[tool.pylint.main]
init-hook = "import sys; from pathlib import Path; import platform; venv_base = Path.cwd() / 'venv' / 'lib'; python_version = f'python{platform.python_version_tuple()[0]}.{platform.python_version_tuple()[1]}'; venv_path = venv_base / python_version / 'site-packages'; sys.path.insert(0, str(venv_path)) if venv_path.exists() else None"

[tool.pylint."messages control"]
disable = [
    "import-error",  # Disable import errors - modules are available in venv
]

